modify

+++ benchmarks/benchmark_triton_impl/benchmark_custom_triton_impl.py
@@ -13,8 +13,8 @@ def test_lightning_attention_implementations(model_params):
 
     batch_size = 2
     seq_len = 64
-    # dtype = torch.bfloat16
-    dtype = torch.float32
+    dtype = torch.bfloat16
+    # dtype = torch.float32
     device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
 
     hidden_states = torch.randn(
